Get lint-test-expectations passing https://bugs.webkit.org/show_bug.cgi?id=228999
Reviewed by Ryan Haddad.

Tools:

* Scripts/webkitpy/layout_tests/controllers/layout_test_finder.py:
(LayoutTestFinder._is_test_file): Added a list of patterns for tests to skip
and ensured that 'boot.xml' and 'root.xml' (spurious files created during
the run of the WebKit1 bot) are on the list.
* Scripts/webkitpy/layout_tests/controllers/layout_test_finder_unittest.py:
(LayoutTestFinderTests.test_is_test_file): Added a test for the changes.

LayoutTests:

* TestExpectations: Remove expectation for non-existent test.
* platform/ios-simulator/TestExpectations: Ditto.
* platform/ios-wk2/TestExpectations: Ditto.
* platform/ios/TestExpectations: Dito.
* platform/mac-wk1/TestExpectations: Ditto. Also, the expectations for the non-test
'boot.xml' and 'root.xml' have been moved to workarounds in webkitpy.
* platform/mac/TestExpectations: Remove references to non-existent tests. Also combined
one expectation that was not linting due to a problem like the one described in bug
120081.
* platform/win/TestExpectations: Ditto.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281010 268f45cc-cd09-0410-ab3c-d52691b4dbfc