REGRESSION(r280382): [GTK] 2.33.3 does not build with gtk-doc enabled, installs broken pkg-config files https://bugs.webkit.org/show_bug.cgi?id=229152
Patch by Michael Catanzaro <mcatanzaro@gnome.org> on 2021-08-18
Reviewed by Philippe Normand.

Source/JavaScriptCore:

* PlatformGTK.cmake:
* javascriptcoregtk.pc.in:

Source/WebKit:

CMake is expanding templates in the pkg-config files that are not supposed to be expanded.
Oops! Let's switch back to using @SVN_REVISION@ instead of ${SVN_REVISION} as the template
for inserting the SVN revision into the pkg-config file, so we can tell CMake to leave the
${} variables alone.

* PlatformGTK.cmake:
* PlatformWPE.cmake:
* Shared/glib/BuildRevision.h.in:
* gtk/webkit2gtk-web-extension.pc.in:
* gtk/webkit2gtk.pc.in:

Tools:

* glib/apply-build-revision-to-files.py:
(main):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281192 268f45cc-cd09-0410-ab3c-d52691b4dbfc