Do not terminate relaunched network process when getting network process connection https://bugs.webkit.org/show_bug.cgi?id=229236
Reviewed by Chris Dumez.

When UI process fails to get network process connection for the first time, it will retry on next runloop
iteration. On the retry, it terminates existing network process and relaunch a network process, because the
failure may indicate something is wrong in the network process. If existing network process is different from
the one on first try, the existing process can be functional and we should not terminate it.

* UIProcess/WebsiteData/WebsiteDataStore.cpp:
(WebKit::WebsiteDataStore::getNetworkProcessConnection):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@281199 268f45cc-cd09-0410-ab3c-d52691b4dbfc